rename interface to server
